task_manager/requirements.txt
boto3
pandas
cvat_sdk
python-dotenv